requirements.txt
django
django-bootstrap5
Pillow